chromium/third_party/blink/web_tests/external/wpt/scroll-animations/css/at-scroll-timeline-before-phase-expected.txt

This is a testharness.js-based test.
FAIL Animation does not apply when timeline phase is before assert_equals: expected "0px" but got "100px"
Harness: the test ran to completion.
